changes in assess environment

+++ assess_environment.h
@@ -2,8 +2,8 @@
 #define ASSESS_ENVIRONMENT_H
 
 
-uint16_t get_line_position(void);
-uint8_t get_line_number(void);
+int get_line_position(void);
+int get_line_number(void);
 void assess_environment_start(void);
 
 #endif /* ASSESS_ENVIRONMENT_H */
